Move TCE content to Want to be Paid page
