+++ app/controllers/payments/WantToBePaidController.scala
@@ -16,49 +16,90 @@
 
 package controllers.payments
 
+import controllers.AnswerExtractor
 import controllers.actions._
 import forms.payments.WantToBePaidFormProvider
-import pages.Waypoints
-import pages.payments.WantToBePaidPage
+import models.Income._
+import models.RelationshipStatus._
+import pages.income.{ApplicantIncomePage, ApplicantOrPartnerIncomePage}
+import pages.payments._
+import pages.{RelationshipStatusPage, Waypoints}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.payments.WantToBePaidView
+import views.html.payments._
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
 class WantToBePaidController @Inject()(
-                                         override val messagesApi: MessagesApi,
-                                         sessionRepository: SessionRepository,
-                                         identify: IdentifierAction,
-                                         getData: DataRetrievalAction,
-                                         requireData: DataRequiredAction,
-                                         formProvider: WantToBePaidFormProvider,
-                                         val controllerComponents: MessagesControllerComponents,
-                                         view: WantToBePaidView
-                                 )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
+                                                override val messagesApi: MessagesApi,
+                                                identify: IdentifierAction,
+                                                getData: DataRetrievalAction,
+                                                requireData: DataRequiredAction,
+                                                val controllerComponents: MessagesControllerComponents,
+                                                coupleUnder50kView: WantToBePaidCoupleUnder50kView,
+                                                coupleUnder60kView: WantToBePaidCoupleUnder60kView,
+                                                coupleOver60kView: WantToBePaidCoupleOver60kView,
+                                                singleUnder50kView: WantToBePaidSingleUnder50kView,
+                                                singleUnder60kView: WantToBePaidSingleUnder60kView,
+                                                singleOver60kView: WantToBePaidSingleOver60kView,
+                                                formProvider: WantToBePaidFormProvider,
+                                                sessionRepository: SessionRepository
+                                              )(implicit ec: ExecutionContext)
+  extends FrontendBaseController
+    with I18nSupport
+    with AnswerExtractor {
 
-  val form = formProvider()
+  private val form = formProvider()
 
   def onPageLoad(waypoints: Waypoints): Action[AnyContent] = (identify andThen getData andThen requireData) {
     implicit request =>
 
       val preparedForm = request.userAnswers.get(WantToBePaidPage) match {
-        case None => form
         case Some(value) => form.fill(value)
+        case None        => form
       }
 
-      Ok(view(preparedForm, waypoints))
+      getAnswer(RelationshipStatusPage) {
+        case Married | Cohabiting =>
+          getAnswer(ApplicantOrPartnerIncomePage) {
+            case BelowLowerThreshold => Ok(coupleUnder50kView(waypoints))
+            case BetweenThresholds   => Ok(coupleUnder60kView(preparedForm, waypoints))
+            case AboveUpperThreshold => Ok(coupleOver60kView(preparedForm, waypoints))
+          }
+
+        case Single | Divorced | Separated | Widowed =>
+          getAnswer(ApplicantIncomePage) {
+            case BelowLowerThreshold => Ok(singleUnder50kView(waypoints))
+            case BetweenThresholds   => Ok(singleUnder60kView(preparedForm, waypoints))
+            case AboveUpperThreshold => Ok(singleOver60kView(preparedForm, waypoints))
+          }
+      }
   }
 
   def onSubmit(waypoints: Waypoints): Action[AnyContent] = (identify andThen getData andThen requireData).async {
     implicit request =>
 
       form.bindFromRequest().fold(
-        formWithErrors =>
-          Future.successful(BadRequest(view(formWithErrors, waypoints))),
+        formWithErrors => {
+          getAnswerAsync(RelationshipStatusPage) {
+            case Married | Cohabiting =>
+              getAnswerAsync(ApplicantOrPartnerIncomePage) {
+                case BelowLowerThreshold => Future.successful(BadRequest(coupleUnder50kView(waypoints)))
+                case BetweenThresholds   => Future.successful(BadRequest(coupleUnder60kView(formWithErrors, waypoints)))
+                case AboveUpperThreshold => Future.successful(BadRequest(coupleOver60kView(formWithErrors, waypoints)))
+              }
+
+            case Single | Divorced | Separated | Widowed =>
+              getAnswerAsync(ApplicantIncomePage) {
+                case BelowLowerThreshold => Future.successful(BadRequest(singleUnder50kView(waypoints)))
+                case BetweenThresholds   => Future.successful(BadRequest(singleUnder60kView(formWithErrors, waypoints)))
+                case AboveUpperThreshold => Future.successful(BadRequest(singleOver60kView(formWithErrors, waypoints)))
+              }
+          }
+        },
 
         value =>
           for {
