+++ app/views/payments/WantToBePaidSingleUnder50kView.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import controllers.income.routes
+@import controllers.payments.routes
 
 @this(
         layout: templates.Layout,
@@ -24,28 +24,28 @@
 
 @(waypoints: Waypoints)(implicit request: Request[_], messages: Messages)
 
-@layout(pageTitle = titleNoForm(messages("taxChargeExplanation.single.under50k.title"))) {
+@layout(pageTitle = titleNoForm(messages("wantToBePaid.single.under50k.title"))) {
 
-    @formHelper(action = routes.TaxChargeExplanationController.onSubmit(waypoints)) {
+    @formHelper(action = routes.WantToBePaidController.onSubmit(waypoints)) {
 
-        <h1 class="govuk-heading-xl">@messages("taxChargeExplanation.single.under50k.heading")</h1>
+        <h1 class="govuk-heading-xl">@messages("wantToBePaid.single.under50k.heading")</h1>
 
         <p class="govuk-body">
-            @messages("taxChargeExplanation.single.under50k.p1")
-            <a href="@messages("taxChargeExplanation.single.under50k.p1.link.href")" class="govuk-link">@messages("taxChargeExplanation.single.under50k.p1.link.text")</a>.
+            @messages("wantToBePaid.single.under50k.p1")
+            <a href="@messages("wantToBePaid.single.under50k.p1.link.href")" class="govuk-link">@messages("wantToBePaid.single.under50k.p1.link.text")</a>.
         </p>
 
         <div class="govuk-warning-text">
             <span class="govuk-warning-text__icon" aria-hidden="true">!</span>
             <strong class="govuk-warning-text__text">
                 <span class="govuk-warning-text__assistive">@messages("site.warning")</span>
-                @messages("taxChargeExplanation.single.under50k.p2.1")
-                <a href="@messages("taxChargeExplanation.single.under50k.p2.link.href")" class="govuk-link">@messages("taxChargeExplanation.single.under50k.p2.link.text")</a>.
-                @messages("taxChargeExplanation.single.under50k.p2.2")
+                @messages("wantToBePaid.single.under50k.p2.1")
+                <a href="@messages("wantToBePaid.single.under50k.p2.link.href")" class="govuk-link">@messages("wantToBePaid.single.under50k.p2.link.text")</a>.
+                @messages("wantToBePaid.single.under50k.p2.2")
             </strong>
         </div>
 
-        <p class="govuk-body govuk-!-margin-bottom-7">@messages("taxChargeExplanation.single.under50k.p3")</p>
+        <p class="govuk-body govuk-!-margin-bottom-7">@messages("wantToBePaid.single.under50k.p3")</p>
 
         <input type="hidden" name="value" value="true" />
